PedantHashComparator normalizes resources in URIs
Our pedant tests frequently conflate the address the server is
listening at (platform.server) with the URI the client should expect
out of the API. This becomes important in cases like Automate/Chef
Server all in 1 installs, where the Automate service proxies requests
to the Chef Server.

This solves the problem by normalizing every appearance of
platform.server into platform.base_resource_url. An alternate solution
would be to update the pedant tests to distinquish the usages. That
would be invasive, changing many tests, and be fragile, every test has
to carefully distinguish the two useages, and we don't typically test
proxied configuration in regular Chef-Server releases.

Normalizing is both elegant and opaque. It is elegant, in that it
confines the change to just a few lines of code, and avoids the need
to be careful. However it's opaque, in that it introduces yet more
abstraction into a codebase which suffers from that already. It also
risks confusion, in that any difference report may contain
'differences' that are actually normalized away.

Signed-off-by: Mark Anderson <mark@chef.io>